+++ src/layout/settings/index.ts
@@ -1,5 +1,7 @@
 import Settings from './Settings.svelte';
+import { settings } from './stores/settings';
 
 export {
-    Settings
+    Settings,
+    settings
 }